consistency: make the argument name of _flip_modes() the same for all ops
Commit a141d2bf "resolve degenerate use of 'mode' in OperatorAdapter
and LinearOperator" was pushed without ensuring consistency with
other operator definitons and correct exception strings. This commit
aims to resolve this.

The argument variable of _flip_modes(…) is now always called 'trafo'
and the Exception thrown for invalid values of it is
ValueError("invalid operator transformation"). To aid readability,
numerical constants were replaced with correspondingly named
constants (ADJ, INV).