modified domain.local_shape's docstring
It is rendered in the docs before domain.shape and as such hard
to understand, because it explains two concepts (shape, MPI data
splitting) at once.

Therefore I added the hint that it is mainly relevant for MPI
and that the docstring of domain.shape explains the shape concept
in general.