Highlight that keys in a multi-domain are sorted
As the ordering of the random numbers within a multi-field are dependent
on the ordering of the keys, it is highly relevant even for the average
nifty user that keys are sorted. Hence, state so in the docstrings of
MultiDomain, MultiField and in the `from_random` method in sugar.py.