Fix complex sample variance
This fix is necessary to make samples consistent with the Hamiltonian. Fixing the variance this way makes np.std(s) of a samples s with variance 1 havyield sqrt(2) and not 1. However, this fix is more compatible with the definition of the complex Gaussian according to wikipedia, and compatible with the definition of the Fisher metric.