correlated_fields.py: Move ZM to finalize
The zero-mode must not be known for constructing amplitudes. Only when
finalizing the model, the zero-mode needs to be specified. Honor this by
pulling all parameters relating to the zero-mode to the `finalize`
method.

Move `moment_slice_to_average` outside of `CorrelatedFieldMaker` as it
does not depend on nor require the inherent structure of the class.