optmize.py: Extend DIY CG interface
To more easily switch between the DIY CG and the one in JAX, make the
DIY implementation accept a starting position `x0` and return a tuple of
the position as well as information on the iteration count akin to JAX
and scipy.

Furthermore, PEPify variable names and format the code.