Add jit member function
Usually everything should be compiled.
The jit member function allows doing so in a very convenient way for Likelihood and StandardHamiltonian.
One still has to compile in the main code because the Likelihood class has no function fgor value_and_Grad, but this is needed and should be compiled as well.
Nevertheless, great speedups can be seen for the categorical demo.