Compare apples to apples in nifty2jax demo
Actually compare the performance of the forward **and** backwards pass
of the model and not just the assembly of the linearization in NIFTy.
Now, there is no difference between JAX and NIFTy in performance for
models with more than \approx 2e+5 number parameters. This is because,
at roughly this point the FFT starts to dominate the computation.