Implement and demo larger refinement filters
Optionally, the edges of the kernel can be dampened thus deviating from
the optimal linear filter and the Gaussian process limit. This scheme
correctly adjusts the excitation correlation matrices. The idea of
down-weighting the edges of the kernel, is avoid imprinting a boxy
structure into a model that otherwise has a homogeneous and isotropic
(i.e. circular) kernel.