re.refine: Fix severe error in distance calc
The distances during refinement were wrongly calculated to always be one
layer ahead, i.e. they were usually half (divided by `_fine_sz`) of what
they were supposed to be. If all of the refinement started off from a
wrong distance this would not have been a problem. However, the very
first layer at zero depth used the correct distances thus poisoning the
refinement.

Introduce an end-to-end test to very that the covariance is correct in
the special case of refining 3 pixels to 2.