OperatorSpectrum: adjust to new scipy interface
After a version bump, scipy probes the operator with a int8 array.
Therefore, we cannot make sure that the input of the Operator is has
floating dtype. As an alterantive dtype check, is is now checked that
all dtypes of a MultiField are the same which is necessary to put all of
its entries into a single flattened array.