SampleList: fix IO bug
Previously, if a SampleList that contains n samples has already been
saved and afterwards a SampleList with n-m (m>0, n-m>0) is saved and
then loaded again, it considered the samples between n-m and n as part
of the new, updated SampleList. This is faulty behaviour.

This patch adds a fix and tests that make sure that this faulty
behaviour does not happen again in the future.